DEV: move storage services to lib
