+++ libs/shared/data-access/storage/src/index.ts
@@ -1,1 +1,3 @@
+export * from './lib/generic-storage.service';
+export * from './lib/local-storage.service';
 export * from './lib/shared-data-access-storage.module';
